Clarify use of `--net host` option in README.
The way dnsmasq does DHCP requires that the device serving the requests
be part of the same subnet. Since docker containers use virtual devices,
one would need to bridge them to the physical device that the physical
computers are connected to make it work. This is what tools like
pipework were originally for and with `--net host` fixes.

I discovered that all other functions of dnsmasq however can be run with
careful opening of tcp and udp ports. So this allows us to only run this
DHCP+tftp+pxe server using `--net host` and leave dns to run in a normal
container.